- update competition without tests; - update water activity;

+++ backend/waveCoach/domain/src/main/kotlin/waveCoach/domain/HeatToInsert.kt
@@ -3,5 +3,5 @@ package waveCoach.domain
 data class HeatToInsert(
     val competitionId: Int,
     val waterActivityId: Int,
-    val score: Float
+    val score: Int
 )
