+++ backend/waveCoach/http/src/main/kotlin/waveCoach/http/model/input/HeatInputModel.kt
@@ -1,6 +1,6 @@
 package waveCoach.http.model.input
 
 data class HeatInputModel(
-    val score: Float,
+    val score: Int,
     val waterActivity: WaterActivityInputModel
 )
